utils/path: tidy&generalize templating resources
This generalizes our access to template files. All resources are gathered
in a dedicated directory under the root package ('res'). We include this
as package data and provide a simple access interface with our utils.path
module.

The Jinja2 PackageLoaders are utilized under the hood. So the same interfaces
can be used directly as well (as is done in utils.wrapping).