experiment: fix a registration erro
Since adding proper type information to Experiment classes, an error was
introduced in the handling of experiment registration during class  definition.

The default value of an experiment's NAME attribute has become '' and is always
defined (This should have been broken even without type information).
This should not cause registration.

An experiment must always have a NAME attribute  attached to it directly, to be
registered as a valid experiment, otherwise we risk overwriting an exisitng
experiment.