revert/adapt: BaseVersionFilter back to ISource
This adapts the previous changes. BaseVersionFilter becomes a descendant
of ISource again. We would have to specify _local/_remote attributes for
a version filter otherwise (attrs).

There is no reason why a 'Source' interface should not know about local
and remote locations. So we just add these as abstract properties to
ISource.

However, a future commit needs to deal with BaseVersionGroup:
There is no sensible way to define a local or remote property for this.

As a simple solution we could do a little amputation and throw
NotImplemented or ValueError when accessing local/remote of a version
group.