reports: remove unfinished 'report' functionality.
This was a quick and dirty implementation to get access to the database
in a convenient fashion. As this pulls in pandas globally for all of
benchbuild, we remove it for now.

This is better suited for a separate project/namespace, so we can avoid
the heavy pandas dependency (or similar ones).