project/tests: require CONTAINER to be filled by the project author.
We do not enforce the CONTAINER attribute to be filled explicitly.

However, benchbuild will fail on every operation that accesses the
member instance attribute 'container' or the 'CONTAINER' ClassVar.

This is intended for now. The plan is to distinguish 'classic' projects
from 'container projects' using mixins/inheritance in a future commit.