provide structural subtypes for sources
This converts the interface <-> ABC relationship of ISource and BaseSource into
more pythonic structural subtyping.

We provide several protocols, which are identical to plain simple ABC's at
runtime:
 * Fetchable - Something that can be downloaded from a remote location
 * Expandable - Something that can decide whether it wants to take part in
                version expansion
     (e.g. a source that provides elements in a tuple of a cross-product)
 * Versions - Something that can select and enumerate versions.

By default, we provide a 'FetchableSource' that can be used as a base class that
supports all 3 protocols. Future commits, will make more fine-grain use of these
protocols. This is a quick-and-dirty 1:1 conversion from the previous state.

Directly related to this change, we add a GitSubmodule type that is basically
identical to a Git source, without taking part in version expansion.