Add support for GitSubmodules (#333)
* provide structural subtypes for sources

This converts the interface <-> ABC relationship of ISource and BaseSource into
more pythonic structural subtyping.

We provide several protocols, which are identical to plain simple ABC's at
runtime:
 * Fetchable - Something that can be downloaded from a remote location
 * Expandable - Something that can decide whether it wants to take part in
                version expansion
     (e.g. a source that provides elements in a tuple of a cross-product)
 * Versions - Something that can select and enumerate versions.

By default, we provide a 'FetchableSource' that can be used as a base class that
supports all 3 protocols. Future commits, will make more fine-grain use of these
protocols. This is a quick-and-dirty 1:1 conversion from the previous state.

Directly related to this change, we add a GitSubmodule type that is basically
identical to a Git source, without taking part in version expansion.

* Add backward compatible support for Protocol to source/base
* Add backward compatible support for Protocol to utils/run
* Change source tests to use FetchableSources
* Test disabled version expansion for GitSubmodule
* Migrate test_project to FetchableSources
* Add lookup for non-expanded sources to project.source_of

This provides convenient access from a project to a given (non-expanded) source as usual.

* Make sure Git sources take care of their submodules during clone.

* Add fixtures to test git repositories

This adds 3 fixtures to our tests for easier testing against git repositories.

1. mk_git_repo:

A factory test fixture that can generate a random git repository in temporary
directory. The repo will be filled with a number of random small commits.
You cannot make any assumptions about file contents.
You can add an optional git submodule to the repository by giving it an
existing git repo, created by this fixture before.

2. simple_repo:
Wrap mk_git_repo and provide a simple git repository without submodule.

3. repo_with_submodule:
Wrap mk_git_repo and create 2 repositories, where one is submodule of the other.

* Add 2 tests for git source fetching

This adds 2 tests:
 1. Make sure a git repo can be fetched
 2. Make sure that a git repo also clones it's submodules

* Add more git source test cases

This Adds more tests for the following behavior:
 1. Submodules can be fetched outside of the main repo.
2. Submodules can be fetched (again) inside of the main repo. This won't change
the main repo.
 3. Repos can list their versions

* Test if a submodule exists twice when cloned outside it's main.

The main git repo will clone recursively. Therefore, we will end up
with two clones of the same submodule. One inside the main repo and one
we have declared in the GitSubmodule.

* Make sure submodules are cached in a flat filepath

Even if a GitSubmodule is placed relative to a Git source, make sure that they
are in separate directories in side 'tmp_dir'.

Repos: A, B (B.local = A/foo/B)

Example:

  tmp_dir:
   A
   A-foo-B

  build_dir
   A
   A/foo/B

* Increase default log level to ERROR

We only have 2 messages of level CRITICAL in benchbuild. Both of which could
easily be of type ERROR.
There is no time during the execution of benchbuild, where we would not
want to see error messages.

* Add edge to edge (e2e) tests for git submodules

Edge to edge tests in benchbuild should avoid relying on the cli
and any real projects to test our API.

This commits adds 2 e2e tests, when we use Git and GitSubmodules:

1. Check, if a projects emits any versions, when Submodules are involved.
2. Check, if we can execution a full benchbuild run. Without touching any
   real compilation/execution commands.

* Test/Fix: Reset benchbuild's config

After each use of the fixture 'project_cls' we need to reset the
global configuration, otherwise other tests may fail.

* Adjust pylint config to remove problems we cannot/won't fix globally.

Those below are debatable, but I am sure it looks cleaner than disabling
these on a file-by-file basis.

The adjusted values are:
 - Accept 'T' as a valid variable name everywhere
 - Maximum number of function arguments to 8, we have contructors that
   require that many.
 - Maximum number of ancestors to 10, we use types with a deep
   hierarchy and would trigger warnings in our own code, which directly
   causes only a fraction of the ancestor-count (typing.Protocol).