fix(project): project registration for groups
This fixes project registration for grouped projects. As soon
as some class attributes were defined on the group level of the type
hierarchy, we wouldn't register any child of this group class.

We fix this by preferring the attributes of the subclass and take the
attributes of the superclass, if we can't find any.