feat(environments): step-wise image construction
This modifies benchbuild's container API to do image construction steps.
The user specifies a high-level API command (CreateImage) and benchbuild
adds the required layers to the container image in our data model.

All layers are tracked with a state: PRESET / ABSENT. The concrete image
repository in use can then make sure that the required layers are spawned
on disk.
An image is completed as soon as all layers are marked as PRESENT.
Any cleanup necessary is only useful, if all layers are marked PRESENT.

On disk we will spawn a build-container that can be put on 'hold' during
layer construction. This makes a container exempt from automatic deletion
at the end of an API command.

Frontend changes should be minimal.
This commit moves a few type declarations around, mostly from the service layer
down into the domain.