feat(environments): make layers hashable
All layers are already immutable. However, the use of dictionaries
makes them not suitable for hashing.

With this patch we convert them to hashable data structures as soon
as the layer is constructed.

All dictionaries used for:
  RunLayer
  UpdateEnv

will be converted to a tuple of tuples.

For example, the dict:
{
  'BB_BUILD_DIR': '/foo/bar'
}
becomes
(('BB_BUILD_DIR', '/foo/bar'),)