feat(environments): add 'rmi' subcommand to delete images.
This adds a new subcommand ``benchbuild container rmi`` to delete a
series of built images.
By default, this removes all experiment images selected by the cli,
e.g.:

``benchbuild container rmi -E raw`` removes *all* images associated with
the 'raw' experiment.

If you specify ``--with-projects`` it will also remove all project
container images. The remove is done with '-f', forcing the delete.

This requires an experiment to sample project versions from. If you want
to delete the bare project images, you still have to provide an
experiment that is able to sample the corresponding project versions, or
benchbuild will not be able to select the images for deletion.