feat(environments): enable debugging of failed image builds
This adds support for a light debug mode. If a layer cannot be created, we will
store the failed state in a configurable container image. From there we will
create a fresh build container and drop into a shell inside that container.

You can then investigate your image state, retry the command or run any
necessary ad-hoc fixes that might be required.

If you leave the debug mode with an exit code of 0, we will commit the image to
the original tag and later commands in the pipeline will proceed as normal.
Use this with caution, because we won't finish building the container where we
left of. You have to complete the image build yourself, if you are in a debug
session.
If you want to leave the debug session and skip the commit, just invalidate the
session by returning with an exit code != 0, e.g., by running 'false' or hitting
Ctrl-C before typing 'exit'.