feat(slurm): support variable number of subcommand arguments
This adds support for a variable number of subcommand arguments. Before,
all arguments provided via ``-S`` switch to the subcommand ``benchbuild
slurm`` have been treated as a single argument.

With this commit, we remove the ``-S`` switch and replace it with a set
of variable arguments that are placed after a double-dash.

Example:

``benchbuild slurm -E raw bzip2 -- container --runroot /foo/bar``

This works for the 2 subcommands: ``run`` and ``container``. You can
mess things up, use with caution.

Fixes #396