fix(settings): BB_ENV is ignored when no config file is loaded"
The ``_INNER_NODE_VALUE`` schema did not cover inner nodes that define
a description as well.

In our default settings, BB_ENV is defined as such an inner node.
This creates a ``value`` entry at BB_ENV that is then fillen with the
dictionary from our default config.

This makes it impossible to fill the key ``BB_ENV`` from the
environment, as we now captured BB_ENV_VALUE as new leaf value.

This fixes #413