feat(workload): Add prototype support for workloads.
This converts our implementation of workloads in the form of 'compile' and
'run_tests' into a generic support for workload functions.

The initial implementation aims for feature parity with our current behavior.
Therefore, we only add a 'Phase' enum to determine the execution time of the
workload.

Workloads are executed in insertion order, which is the order of definition in
the project.

A new workload is added by decorating a function with the 'phase' class
function.

This will be a breaking change.