feat(actions): support partial revisions in SetProjectVersion
This commit now supports partial revisions in SetProjectVersion.
We match the provided revision strings with the source revisions in the
associated project.
The existing active variant will be updated with the new partial variant
context.

This suffers from a version collision problem:
If a revision string can be found in more than one source, we will
update all variants. Right now we have no clean way of resolving it.
For now we keep the behavior because it should(!) not happen for git
repositories.