feat(jobs): add RunJob and RunJobs actions
This adds a minimal example for our new JOBs interface. As a project
designer you can now declare run-time jobs of a project.
The added classes RunJob and RunJobs are actions that can be used to
schedule all jobs for execution.

This replaces the old Run action.

Signal handling and exceptions should work as usual. Testing needs to be
added.