fix(jobs): allow jobs to run as wrapped binaries
This commit still needs some work afterwards. We have to avoid wrapping
a binary twice because the sequence is as follows:

```
original
wrap(original) -> mv original original.bin -> create wrapper original
```

After the first wrap we have our wrapper in original and the original
binary in original.bin.
Repeated calls to wrap for the same binary will result in an infinite
loop, because we mv our wrapper to original.bin and create a new
wrapper in original.