feat(wrapping): avoid wrapping the same command twice.
This changes benchbuild's wrapping behavior. ``wrap`` and
``wrap_dynamic`` store wrapped paths in the companion object (Normally
a ``benchbuild.project.Project``).

If a path has already been wrapped, we do not move it to the binary
path. However, we still create a fresh wrapper script at the original
location of the binary.

User-facing behavior should be unchanged.