feat(command): add support for WorkloadSet.
This adds a WorkloadSet that can be used as a key in a project's job
mapping.

benchbuild/bzip2 serves as a minimal example for this feature.
The JOBS ClassVar has to map a WorkloadSet to a list of jobs.

An experiment can make use of custom actions to filter actions based on
filter conditions.

See ``RunJobs`` in ``benchbuild/utils/actions.py`` for an example.