feat(command): add example `OnlyIn` for WorkloadSet customization.
This customization returns the captured WorkloadSet only if the given
project is part of the given RevisionRange.

This commit serves as an example code and makes the following
assumption:

The source of the project is already checked out in the build directory(!).
Other uses of revision ranges operate on BenchBuild's source cache.