feat(command): make use of rendered and unrenderer PathToken explicit.
This makes the protocol for PathToken explicit in the sense that we
now require dedicated properties and methods for the unrendered and
rendered version of the token.

If a rendered version of a token is requested by the user or the API
that cannot be provided yet, an error is logged and the unrendered
version will be returned.

For tokens that require information about a project, make sure to
wrap your existing Command with a ProjectCommand.