feat(source): introduce type skeleton and prototypes for context-aware version enumeration.
This introduces a (currently) non-working stage of context-aware version enumeration.

The main idea will be a 2-phase design for version enumeration:

* Phase 1: Enumerate all context-free sources, i.e., sources that do not need to know anything about the context they are being enumerated in.

* Phase 2: Iterate over all versions from Phase 1 and feed each resulting variant context into all context-sensitive sources.
These may provide a new list of versions for each context they are given.

The resulting list of variant context objects will be iterated on by benchbuild just as before.

Integration for CLI is pending. In essence, all sampling and filtering should be possible on the new enumeration just as before.