performance optimization: add config option to disable the database
Before, we always used the sqlite in-memory database to "disable" the
database. Now, there is a proper configuration option for this. The
advantage is that this can lead to significant performance improvements
when benchbuild is invoked often, for example, in ./configure based
projects. These kinds of projects invoke the wrapped compiler often to
determine whether certain feature are available.